[ Refactor ] Route이름 수정 (Prepare & Setting)
