+++ src/assets/scss/layouts/prepare/LevelBoxTop.scss
