+++ src/assets/scss/layouts/prepare/NicknameInputBox.scss
