+++ src/assets/scss/pages/Prepare.scss
@@ -1,4 +1,4 @@
-.gameprepare {
+.prepare {
   .button--bottom {
     position: fixed;
     width: 100%;
