+++ src/assets/scss/pages/Setting.scss
@@ -1,6 +1,6 @@
 @import 'utils/variables.scss';
 
-.gamesetting {
+.setting {
   display: flex;
   flex-direction: column;
   align-items: center;
